refactor: move position of PodcastController
